+++ Source/DirectiveUtilitiesBlueprintNodes/Public/DirectiveUtilitiesBlueprintNodes.h
@@ -1,0 +1,20 @@
+// Copyright (c) 2026 Unreal Directive. Licensed under the MIT License.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Modules/ModuleManager.h"
+
+class UObject;
+
+class FDirectiveUtilitiesBlueprintNodesModule : public IModuleInterface
+{
+public:
+	virtual void StartupModule() override;
+	virtual void ShutdownModule() override;
+
+private:
+	void HandleAssetLoaded(UObject* Asset);
+
+	FDelegateHandle AssetLoadedHandle;
+};