+++ Source/DirectiveUtilitiesBlueprintNodes/Public/Nodes/DirectiveUtilMapNodeMigration.h
@@ -1,0 +1,10 @@
+// Copyright (c) 2026 Unreal Directive. Licensed under the MIT License.
+
+#pragma once
+
+class UBlueprint;
+
+namespace DirectiveUtilMapNodeMigration
+{
+	DIRECTIVEUTILITIESBLUEPRINTNODES_API bool UpgradeLegacyAppendNodes(UBlueprint& Blueprint);
+}